+++ services/autoscaling/tests/unit/test_utils_cluster_scaling.py
@@ -19,14 +19,12 @@ from models_library.generated_models.docker_rest_api import Node as DockerNode
 from pydantic import TypeAdapter
 from pytest_simcore.helpers.monkeypatch_envs import setenvs_from_dict
 from pytest_simcore.helpers.typing_env import EnvVarsDict
-from simcore_service_autoscaling.core.errors import Ec2InvalidDnsNameError
 from simcore_service_autoscaling.core.settings import ApplicationSettings
 from simcore_service_autoscaling.models import AssociatedInstance, EC2InstanceData
 from simcore_service_autoscaling.utils.cluster_scaling import (
     associate_ec2_instances_with_nodes,
     ec2_startup_script,
     get_machine_buffer_type,
-    node_host_name_from_ec2_private_dns,
     sort_drained_nodes,
 )
 from simcore_service_autoscaling.utils.utils_docker import (
@@ -50,34 +48,6 @@ def node(faker: Faker) -> Callable[..., DockerNode]:
         )
 
     return _creator
-
-
-@pytest.mark.parametrize(
-    "aws_private_dns, expected_host_name",
-    [
-        ("ip-10-12-32-3.internal-data", "ip-10-12-32-3"),
-        ("ip-10-12-32-32.internal-data", "ip-10-12-32-32"),
-        ("ip-10-0-3-129.internal-data", "ip-10-0-3-129"),
-        ("ip-10-0-3-12.internal-data", "ip-10-0-3-12"),
-    ],
-)
-def test_node_host_name_from_ec2_private_dns(
-    fake_ec2_instance_data: Callable[..., EC2InstanceData],
-    aws_private_dns: str,
-    expected_host_name: str,
-):
-    instance = fake_ec2_instance_data(
-        aws_private_dns=aws_private_dns,
-    )
-    assert node_host_name_from_ec2_private_dns(instance) == expected_host_name
-
-
-def test_node_host_name_from_ec2_private_dns_raises_with_invalid_name(
-    fake_ec2_instance_data: Callable[..., EC2InstanceData], faker: Faker
-):
-    instance = fake_ec2_instance_data(aws_private_dns=faker.name())
-    with pytest.raises(Ec2InvalidDnsNameError):
-        node_host_name_from_ec2_private_dns(instance)
 
 
 @pytest.mark.parametrize("valid_ec2_dns", [True, False])
